# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillSync AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs and internships
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [job, internship, freelance, hackathon]
        - name: search
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/apply:
    post:
      operationId: applyJob
      tags: [jobs]
      summary: Apply for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyJobBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
...
  /applications:
    get:
      operationId: listApplications
      tags: [jobs]
      summary: List applications for current user or recruiter
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobApplication"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [jobs]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"

  # ASSESSMENTS
...
  /freelance/projects:
    get:
      operationId: listFreelanceProjects
      tags: [freelance]
      summary: List freelance projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelanceProjectsListResponse"
    post:
      operationId: createFreelanceProject
      tags: [freelance]
      summary: Post a freelance project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFreelanceProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelanceProject"
  /freelance/projects/{id}:
    get:
      operationId: getFreelanceProject
      tags: [freelance]
      summary: Get freelance project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelanceProject"
  /freelance/projects/{id}/bids:
    get:
      operationId: listBids
      tags: [freelance]
      summary: List bids on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: placeBid
      tags: [freelance]
      summary: Place a bid on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidBody"
      responses:
        "201":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"

  # COLLEGE
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, recruiter, faculty, freelancer_client, admin]
        avatar:
          type: string
        bio:
          type: string
        institution:
          type: string
        location:
          type: string
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, role]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        type:
          type: string
          enum: [job, internship, freelance, hackathon]
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        salary:
          type: string
        deadline:
          type: string
        postedById:
          type: integer
        applicantCount:
          type: integer
        createdAt:
          type: string
        status:
          type: string
          enum: [open, closed]
      required: [id, title, company, type, description, skills, status]
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        type:
          type: string
          enum: [job, internship, freelance, hackathon]
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        salary:
          type: string
        deadline:
          type: string
      required: [title, company, type, description, skills]
    JobsListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [jobs, total, page, limit]
...
    ApplyJobBody:
      type: object
      properties:
        coverLetter:
          type: string
        resumeId:
          type: integer
      required: []
    JobApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, reviewed, shortlisted, rejected, hired]
        coverLetter:
          type: string
        appliedAt:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        user:
          $ref: "#/components/schemas/User"
      required: [id, jobId, userId, status, appliedAt]
    UpdateApplicationBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, shortlisted, rejected, hired]
      required: [status]

    # ASSESSMENTS
...
    FreelanceProject:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        budget:
          type: string
        skills:
          type: array
          items:
            type: string
        deadline:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        clientId:
          type: integer
        bidCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, budget, skills, status, clientId]
    CreateFreelanceProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        budget:
          type: string
        skills:
          type: array
          items:
            type: string
        deadline:
          type: string
      required: [title, description, budget, skills]
    FreelanceProjectsListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/FreelanceProject"
        total:
          type: integer
        page:
          type: integer
      required: [projects, total, page]
    Bid:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        freelancerId:
          type: integer
        amount:
          type: string
        proposal:
          type: string
        deliveryTime:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
        freelancer:
          $ref: "#/components/schemas/User"
      required: [id, projectId, freelancerId, amount, proposal, status, createdAt]
    PlaceBidBody:
      type: object
      properties:
        amount:
          type: string
        proposal:
          type: string
        deliveryTime:
          type: string
      required: [amount, proposal]

    # COLLEGE
...